improved solution testing
* Added test control parameters introduced for hackathon (test_skip etc.)
* Refactor to enable simpler tests per solution, located *in* the solution directory; moved tests/text_excel_integration.py to tools/expected_result_tester.py
* Prototype with afforestation solution only
* Removed top-level tests directory as it is no longer needed (other contents were out of date)
* Cleanup of some excel utils
