--- solution/afforestation/testdata/expected.zip
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a651c002d848adaf6bbf54e993bb873bda97c335b0627f9bcf54ae55c7b1cd42
-size 5811940
