--- tests/.gitignore
@@ -1,6 +1,0 @@
-venv
-.history
-.pytest_cache
-__pycache__
-*.pyc
-*.pyo
