+++ tools/solution_xls_extract.py
@@ -29,7 +29,7 @@ import pandas as pd
 import pytest
 from . import rrs
 
-from tools.util import convert_bool, xls, xli, xln, co, cell_to_offsets
+from tools.util import convert_bool, xls, xli, xln, co
 from tools.vma_xls_extract import VMAReader
 from model import advanced_controls as ac
 
@@ -626,7 +626,7 @@ def write_tam(f, wb, outputdir):
                                       outputdir=tamoutputdir, prefix='tam_pds_')
     if recursive_keys(pds_sources) == recursive_keys(rrs.energy_tam_1_pds_data_sources):
         # the source names are the same for energy_tam_1 & 2, distinguish them here.
-        plausible_2060 = float(tm_tab.cell(*cell_to_offsets('L152')).value)
+        plausible_2060 = xln(tm_tab, 'L152')
         if plausible_2060 == pytest.approx(54539.190092617995):
             arg_pds = 'rrs.energy_tam_2_pds_data_sources'
         elif plausible_2060 == pytest.approx(60153.728317538):
